+++ agent-api/src/test/java/com/intergral/deep/agent/api/spi/IDeepPluginTest.java
@@ -15,25 +15,20 @@
  *     along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.intergral.deep.agent.resource;
+package com.intergral.deep.agent.api.spi;
 
-import com.intergral.deep.agent.api.resource.Resource;
-import com.intergral.deep.agent.api.settings.ISettings;
-import com.intergral.deep.agent.api.spi.IDeepPlugin;
-import com.intergral.deep.agent.api.spi.ResourceProvider;
-import java.util.Collections;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertSame;
 
-/**
- * A resource provider that detects the hava version to add to the resource.
- */
-public class JavaResourceDetector implements IDeepPlugin, ResourceProvider {
+import org.junit.jupiter.api.Test;
 
-  @Override
-  public Resource createResource(final ISettings settings) {
-    final String property = System.getProperty("java.version");
-    if (property == null) {
-      return null;
-    }
-    return Resource.create(Collections.singletonMap("java_version", property));
+class IDeepPluginTest {
+
+  @Test
+  void configure() {
+    final IDeepPlugin iDeepPlugin = new IDeepPlugin() {
+    };
+    assertNotNull(iDeepPlugin.configure(null, null));
+    assertSame(iDeepPlugin, iDeepPlugin.configure(null, null));
   }
 }