+++ examples/dynamic-load/src/main/java/com/intergral/deep/examples/Main.java
@@ -21,7 +21,6 @@ package com.intergral.deep.examples;
 import com.intergral.deep.Deep;
 import com.intergral.deep.agent.api.IDeep;
 import com.intergral.deep.agent.api.reflection.IReflection;
-import com.intergral.deep.agent.api.resource.Resource;
 import com.intergral.deep.agent.api.settings.ISettings;
 import com.intergral.deep.api.DeepAPI;
 import java.nio.file.Path;
@@ -48,6 +47,7 @@ public class Main {
   public static void main(String[] args) throws Throwable {
     // this is only needed in this example as we are using a local built module
     // if using the dependency from maven you do not need to set the path
+    //noinspection DataFlowIssue
     Path jarPath = Paths.get(Main.class.getResource("/").toURI())
         .getParent()
         .getParent()
@@ -70,27 +70,22 @@ public class Main {
     System.out.println(DeepAPI.api().getVersion());
     System.out.println(DeepAPI.reflection());
 
-    // Use the API to register a plugin
-    // This plugin will attach the attribute 'example' to the created snapshot
-    // you should also see the log line 'custom plugin' when you run this example
-    DeepAPI.api().registerPlugin((settings, snapshot) -> {
-      System.out.println("custom plugin");
-      return Resource.create(Collections.singletonMap("example", "dynamic_load"));
-    });
-
     // USe the API to create a tracepoint that will fire forever
     DeepAPI.api()
         .registerTracepoint("com/intergral/deep/examples/SimpleTest", 46,
             Collections.singletonMap("fire_count", "-1"), Collections.emptyList(), Collections.emptyList());
 
     final SimpleTest ts = new SimpleTest("This is a test", 2);
+    //noinspection InfiniteLoopStatement
     for (; ; ) {
       try {
         ts.message(ts.newId());
       } catch (Exception e) {
+        //noinspection CallToPrintStackTrace
         e.printStackTrace();
       }
 
+      //noinspection BusyWait
       Thread.sleep(1000);
     }